Tests/LibWeb/Text/expected/css/slot-style-inheritance.txt

1: rgb(255, 0, 0)
2: rgb(0, 128, 0)
3a: rgb(255, 0, 0)
3b: rgb(0, 0, 255)
4: rgb(255, 0, 0)
5: rgb(0, 0, 255)
6-color: rgb(0, 128, 0)
6-font: italic
7-before: rgb(0, 0, 0)
7-after: rgb(255, 0, 0)
8a: rgb(0, 128, 0)
8b: rgb(0, 128, 0)
8c: rgb(0, 128, 0)
9-before: rgb(0, 0, 0)
9-slot: rgb(0, 0, 0)
9-after: rgb(0, 0, 0)
10-before: rgb(0, 0, 0)
10-after: rgb(0, 128, 0)
11-before: rgb(0, 0, 0)
11-after: rgb(0, 128, 0)
11-removed: rgb(0, 0, 0)
12-before: rgb(255, 0, 0)
12-after: rgb(0, 0, 255)
